Define our own global constants file
Pymoca is not yet able to parse the entire Modelica library, and it
would also not (yet) be easy to distribute and install it along with
RTC-Tools (e.g. with pip). We therefore define our own global constants
file with a subset of the values of Modelica.Constants that we currently
use.

There is also an epsilon constant, which is used to avoid numerical
errors (e.g. division by zero).