Rename Pump to DischargeControlledStructure
The Pump object was often abused to model e.g. discharge-controlled weirs
as well as turbines. The component name of "Pump" was then a bit
misleading. We suggest users choose the more generically named component
added in this commit from here on out.

Pump turns into an alias for DischargeControlledStructure (for backwards
compatibility reasons).

This change applies both to Hydraulic and SimpleRouting libraries.