PartialMuskingum: Fix non-linearity in formulation
Fix the formulation of the Muskingum routing. We make the equation
linear by defining K and x as parameters.

Note that in some more experimental formulations these two parameters
can be time (read: discharge) dependent. If this ever becomes relevant,
we can change PartialMuskingum again. We do not have to worry about
backwards compatibility, as this is an internal model (only the
Muskingum model is public).

Closes #1011