SimpleRouting: Improve scaling by adding Q_nominal to all components
This commit introduces the ability to scale equations in SimpleRouting
models, which is particularly useful for models with discharges much
smaller than 1, ensuring solver accuracy. It also benefits models with a
wide range of discharges, enabling equal relative accuracy for both
large and small discharges.

The new feature is backwards compatible and does not alter existing
behavior. Users should avoid manually setting the nominal of variables
(e.g. QOut.Q) when instantiating a component, but instead use Q_nominal.
Note that this is consistent with the existing practice in Channel Flow.

Also note that equations like `Q / Q_nominal = Q_control / Q_nominal`
will still resolve Q to be an alias of Q_control with the default
settings of RTC-Tools for Pymoca. Scaling to these types of (aliasing)
equations is added for consistency, and for when the user wants to
disable alias detection by Pymoca.