branches: add new version of KNNonlinear
Add a new version of KNNonlinear called KNNonlinearVForm,
which uses the volume instead of q_out_prev and implicit_step_size
to describe the equations.
The volume V is approximated
so that is still differentiable with respect to QOut when QOut = 0.

This V formulation seems to be more stable
when solving a system of equations for implicit timestepping.